requirements.txt
playwright
PyYAML 